routes: move well-known to module, add nodeinfo endpoint
