+++ src/routes/mod.rs
@@ -1,2 +1,3 @@
 pub mod ap;
 pub mod ui;
+pub mod well_known;
